[cmp] propagate decompression errors from bled
If, for example, you have a truncated gz-compressed file and try to
write it to disk, bled_uncompress_with_handles() will return an error.
Previously, this was not reported back to the user -- there's no
indication of any problem. If they're lucky, booting the
partially-written disk will fail in an obvious way, but better to report
an error sooner, even if we cannot give details.